Update softirq-Add-preemptible-softirq.patch and powerpc-preempt-lazy-support.patch
The v5.4-rt version of softirq-Add-preemptible-softirq.patch
containts rcu-Acquire-RCU-lock-when-disabling-BHs.patch.
Therefore, drop the later patch.

Because the updated version of softirq-Add-preemptible-softirq.patch
also modifies the linux/interrupt.h header we update
powerpc-preempt-lazy-support.patch which merged the changes in
powerpc-Fixup-compile-and-lazy-preempt.patch. Drop this patch
too.

suse-commit: 8f4b8c7e351f9a8395fd65c36b3f04da56f37550

+++ include/linux/preempt.h
@@ -100,7 +100,7 @@
 				   (NMI_MASK | HARDIRQ_MASK | SOFTIRQ_OFFSET)))
 #ifdef CONFIG_PREEMPT_RT
 
-#define softirq_count()		(current->softirq_count)
+#define softirq_count()		((long)current->softirq_count)
 #define in_softirq()		(softirq_count())
 #define in_serving_softirq()	(current->softirq_count & SOFTIRQ_OFFSET)
